Login and register endpoint complete. Storage interface needs work
The goal is to have marking sessions initiated before the login to transfer to
the logged in user right after the login. The way the server is
doing that currently doesn't allow for this.

The attempt to have the same schema for logged in user and temporary session is
a complete disaster. default_retrieve and default_commit should be removed