Disable grade entry line protection
The protection is broken. It only works with backspace delete. If the user
tries to delete the whole entry line or press the delete key, it will mess up
the grade entry extraction.

Ace doesn't have support for cancelling an event, but coremirror does.
Codemirror also seems to be more well documented.